WIP checkpoint on reorganization.
