+++ tensorizer/tensorizer.py
@@ -12,52 +12,15 @@ except ImportError:
     resource = None
 
 from enum import Enum
-import subprocess
 import io
-from io import SEEK_SET, SEEK_END
+import tensorizer.stream_io as stream_io
 import numpy
 import struct
 import torch
 import typing
-import time
 import logging
-import sys
-import pathlib
-import os
-import requests
-import argparse
-
-os.environ[
-    "TRANSFORMERS_VERBOSITY"
-] = "error"  # disable missing keys and unexpected key warnings
-
-thisPath = str(pathlib.Path(__file__).parent.resolve())
-sys.path.append(thisPath + "/../../transformers/src")
-sys.path.append(thisPath + "/../../")
-
-from transformers import (
-    AutoTokenizer,
-    AutoModelForCausalLM,
-    AutoConfig,
-    CLIPTextModel,
-    CLIPTokenizer,
-    CLIPTextConfig,
-)
-from transformers.modeling_utils import no_init_weights, PreTrainedModel
-from tokenizers import Tokenizer
-
-from diffusers import (
-    AutoencoderKL,
-    UNet2DConditionModel,
-    StableDiffusionPipeline,
-    LMSDiscreteScheduler,
-)
-from diffusers.modeling_utils import ModelMixin
-from diffusers.configuration_utils import ConfigMixin
-
-import json
-from urllib import request
 import tempfile
+import utils
 
 from collections import OrderedDict
 from typing import Optional, Tuple, Union, List, Iterator, Callable, Dict, Any
@@ -66,90 +29,13 @@ lz4 = None
 
 # Setup logger
 logger = logging.getLogger(__name__)
-logger.setLevel(level=logging.INFO)
-fh = logging.StreamHandler()
-fh_formatter = logging.Formatter(
-    "%(asctime)s %(levelname)s %(filename)s(%(process)d) - %(message)s"
-)
-fh.setFormatter(fh_formatter)
-logger.addHandler(fh)
 
-# =============================================================================
-# From `pipe(7)` manpage:
-#
-# Pipe capacity
-# A pipe has a limited capacity. If the pipe is full, then a write(2) will
-# block or fail, depending on whether the O_NONBLOCK flag is set (see below).
-# Different implementations have different limits for the pipe capacity.
-#
-# Applications should not rely on a particular capacity: an application should
-# be designed so that a reading process consumes data as soon as it is
-# available, so that a writing process does not remain blocked.
-#
-# In Linux versions before 2.6.11, the capacity of a pipe was the same as the
-# system page size (e.g., 4096 bytes on i386). Since Linux 2.6.11, the pipe
-# capacity is 16 pages (i.e., 65,536 bytes in a system with a page size of
-# 4096 bytes). Since Linux 2.6.35, the default pipe capacity is 16 pages, but
-# the capacity can be queried and set using the fcntl(2) F_GETPIPE_SZ and
-# F_SETPIPE_SZ operations. See fcntl(2) for more information.
-#
-# =============================================================================
-# From `fcntl(2)` manpage:
-#
-# Changing the capacity of a pipe
-#
-# F_SETPIPE_SZ (int; since Linux 2.6.35)
-# Change the capacity of the pipe referred to by fd to be at least arg bytes.
-# An unprivileged process can adjust the pipe capacity to any value between the
-# system page size and the limit defined in /proc/sys/fs/pipe−max−size
-# (see proc(5)). Attempts to set the pipe capacity below the page size are
-# silently rounded up to the page size. Attempts by an unprivileged process to
-# set the pipe capacity above the limit in /proc/sys/fs/pipe−max−size yield the
-# error EPERM; a privileged process (CAP_SYS_RESOURCE) can override the limit.
-#
-# When allocating the buffer for the pipe, the kernel may use a capacity larger
-# than arg, if that is convenient for the implementation. (In the current
-# implementation, the allocation is the next higher power-of-two page-size
-# multiple of the requested size.) The actual capacity (in bytes) that is set
-# is returned as the function result.
-#
-# Attempting to set the pipe capacity smaller than the amount of buffer space
-# currently used to store data produces the error EBUSY.
-#
-# Note that because of the way the pages of the pipe buffer are employed when
-# data is written to the pipe, the number of bytes that can be written may be
-# less than the nominal size, depending on the size of the writes.
-#
-# F_GETPIPE_SZ (void; since Linux 2.6.35)
-# Return (as the function result) the capacity of the pipe referred to by fd.
-#
-# =============================================================================
-# Constant for `F_SETPIPE_SZ`, as python3's `fcntl` module doesn't have this
-# defined -- despite the documentation saying that they're there.
-#
-# TODO: Make this work or fail gracefully on non-Linux systems. Not sure if
-#       this is really relevant, as I don't even know if CUDA is available on
-#       non-Linux systems in a production sense.
-F_SETPIPE_SZ = 1031
 
 # Whether the tensor is a parameter or a buffer on the model.
 class TensorType(Enum):
     PARAM = int(0)
     BUFFER = int(1)
     STATEDICT = int(2)
-
-
-# Silly function to convert to human bytes
-def convert_bytes(num):
-    """
-    this function will convert bytes to MB.... GB... etc
-    """
-    step_unit = 1000.0
-
-    for x in ["bytes", "KB", "MB", "GB", "TB"]:
-        if num < step_unit:
-            return "%3.1f %s" % (num, x)
-        num /= step_unit
 
 
 def no_init_or_tensor(loading_code):
@@ -163,257 +49,21 @@ def no_init_or_tensor(loading_code):
         mod.reset_parameters = dummy
     original_empty = torch.empty
 
-    # when torch.empty is called, make it map to meta device by replacing the device in kwargs.
+    # when torch.empty is called, make it map to meta device by replacing the device in
+    # kwargs.
     torch.empty = lambda *args, **kwargs: original_empty(
         *args, **{**kwargs, "device": "meta"}
     )
 
-    with no_init_weights():
-        result = loading_code()
+    result = loading_code()
     for mod in modules:
         mod.reset_parameters = original[mod]
     torch.empty = original_empty
 
     return result
 
-def get_device() -> torch.device:
-    return torch.device("cuda" if torch.cuda.is_available() else "cpu")
 
-
-class CURLStreamFile(object):
-    """
-    CURLStreamFile implements a file-like object around an HTTP download, the
-    intention being to not buffer more than we have to.
-    """
-
-    def __init__(self, uri: str) -> None:
-        # NOTE: `256mb` buffer on the python IO object.
-        self._curl = subprocess.Popen(
-            [
-                "/usr/bin/curl",
-                "--header",
-                "Accept-Encoding: identity",
-                "-s",
-                uri,
-            ],
-            stdout=subprocess.PIPE,
-            bufsize=256 * 1024 * 1024,
-        )
-        # Read our max-fd-size, fall back to 1mb if invalid.
-        pipe_buf_sz = 1024 * 1024
-        try:
-            pipe_file = open("/proc/sys/fs/pipe-max-size", "r")
-            pipe_buf_sz = int(pipe_file.read())
-            logger.debug(f"pipe-max-size: {pipe_buf_sz}")
-        except IOError as e:
-            logger.warning(
-                f"Could not read /proc/sys/fs/pipe-max-size: {e.strerror}"
-            )
-        try:
-            fcntl.fcntl(self._curl.stdout.fileno(), F_SETPIPE_SZ, pipe_buf_sz)
-        except PermissionError as e:
-            logger.warning(
-                f"Couldn't fcntl F_SETPIPE_SZ to {pipe_buf_sz}: {e.strerror}"
-            )
-        self._curr = 0
-        self.closed = False
-
-    def _read_until(
-        self, goal_position: int, ba: Union[bytearray, None] = None
-    ) -> Union[bytes, int]:
-        if ba is None:
-            rq_sz = goal_position - self._curr
-            ret_buff = self._curl.stdout.read(rq_sz)
-            ret_buff_sz = len(ret_buff)
-        else:
-            rq_sz = len(ba)
-            ret_buff_sz = self._curl.stdout.readinto(ba)
-            ret_buff = ba
-        if ret_buff_sz != rq_sz:
-            self.closed = True
-            err = self._curl.stderr.read()
-            self._curl.terminate()
-            if self._curl.returncode != 0:
-                raise (IOError(f"curl error: {self._curl.returncode}, {err}"))
-            else:
-                raise (IOError(f"Requested {rq_sz} != {ret_buff_sz}"))
-        self._curr += ret_buff_sz
-        if ba is None:
-            return ret_buff
-        else:
-            return ret_buff_sz
-
-    def tell(self) -> int:
-        return self._curr
-
-    def readinto(self, ba: bytearray) -> int:
-        goal_position = self._curr + len(ba)
-        return self._read_until(goal_position, ba)
-
-    def read(self, size=None) -> bytes:
-        if self.closed:
-            raise (IOError("CURLStreamFile closed."))
-        if size is None:
-            return self._curl.stdout.read()
-        goal_position = self._curr + size
-        return self._read_until(goal_position)
-
-    @staticmethod
-    def writable() -> bool:
-        return False
-
-    @staticmethod
-    def fileno() -> int:
-        return -1
-
-    def close(self):
-        self.closed = True
-        self._curl.terminate()
-
-    def readline(self):
-        raise Exception("Unimplemented")
-
-    """
-    This seek() implementation is effectively a no-op, and will throw an
-    exception for anything other than a seek to the current position.
-    """
-
-    def seek(self, position, whence=SEEK_SET):
-        if position == self._curr:
-            return
-        if whence == SEEK_END:
-            raise (Exception("Unsupported `whence`"))
-        else:
-            raise (Exception("Seeking is unsupported"))
-
-
-class RequestsStreamFile(object):
-    """
-    RequestsStreamFile implements a file-like object around an HTTP download, the
-    intention being to not buffer more than we have to. Not as fast or efficient
-    as CURLStreamFile, but it works on Windows.
-    """
-
-    def __init__(self, uri: str) -> None:
-        self._uri = uri
-        self._curr = 0
-        self._r = requests.get(uri, stream=True)
-        self.closed = False
-
-    def _read_until(
-        self, goal_position: int, ba: Union[bytearray, None] = None
-    ) -> Union[bytes, int]:
-        if ba is None:
-            rq_sz = goal_position - self._curr
-            ret_buff = self._r.raw.read(rq_sz)
-            ret_buff_sz = len(ret_buff)
-        else:
-            rq_sz = len(ba)
-            ret_buff_sz = self._r.raw.readinto(ba)
-            ret_buff = ba
-        if ret_buff_sz != rq_sz:
-            self.closed = True
-            raise (IOError(f"Requested {rq_sz} != {ret_buff_sz}"))
-        self._curr += ret_buff_sz
-        if ba is None:
-            return ret_buff
-        else:
-            return ret_buff_sz
-
-    def tell(self) -> int:
-        return self._curr
-
-    def readinto(self, ba: bytearray) -> int:
-        goal_position = self._curr + len(ba)
-        return self._read_until(goal_position, ba)
-
-    def read(self, size=None) -> bytes:
-        if self.closed:
-            raise (IOError("RequestsStreamFile closed."))
-        if size is None:
-            return self._r.raw.read()
-        goal_position = self._curr + size
-        return self._read_until(goal_position)
-
-    @staticmethod
-    def writable() -> bool:
-        return False
-
-    @staticmethod
-    def fileno() -> int:
-        return -1
-
-    def close(self):
-        self.closed = True
-        self._r.close()
-        del self._r
-
-    def readline(self):
-        raise Exception("Unimplemented")
-
-    """
-    This seek() implementation is effectively a no-op, and will throw an
-    exception for anything other than a seek to the current position.
-    """
-
-    def seek(self, position, whence=SEEK_SET):
-        if position == self._curr:
-            return
-        if whence == SEEK_END:
-            raise (Exception("Unsupported `whence`"))
-        else:
-            raise (Exception("Seeking is unsupported"))
-
-
-def save_to_state_dict(m, destination, prefix, keep_vars):
-    r"""Saves module state to `destination` dictionary, containing a state
-    of the module, but not its descendants. This is called on every
-    submodule in :meth:`~torch.nn.Module.state_dict`.
-
-    In rare cases, subclasses can achieve class-specific behavior by
-    overriding this method with custom logic.
-
-    Args:
-        destination (dict): a dict where state will be stored
-        prefix (str): the prefix for parameters and buffers used in this
-            module
-    """
-    for name, param in m._parameters.items():
-        destination[prefix + name] = param if keep_vars else param.detach()
-    for name, buf in m._buffers.items():
-        if name not in m._non_persistent_buffers_set:
-            destination[prefix + name] = buf if keep_vars else buf.detach()
-    extra_state_key = prefix + "_extra_state"
-    if (
-        getattr(m.__class__, "get_extra_state", torch.nn.Module.get_extra_state)
-        is not torch.nn.Module.get_extra_state
-    ):
-        destination[extra_state_key] = m.get_extra_state()
-
-
-def state_dict(
-    m: torch.nn.Module, destination=None, prefix="", keep_vars=True
-) -> OrderedDict:
-    if destination is None:
-        destination = OrderedDict()
-        destination._metadata = OrderedDict()
-    destination._metadata[prefix[:-1]] = local_metadata = dict(
-        version=m._version
-    )
-    save_to_state_dict(m, destination, prefix, keep_vars)
-    # m._save_to_state_dict(destination, prefix, keep_vars)
-    for name, module in m._modules.items():
-        state_dict(
-            module, destination, prefix + name + ".", keep_vars=keep_vars
-        )
-    for hook in m._state_dict_hooks.values():
-        hook_result = hook(m, destination, prefix, local_metadata)
-        if hook_result is not None:
-            destination = hook_result
-    return destination
-
-
-class GooseTensorizer:
+class Tensorizer:
     """
     Given a file-like object, either for read or write, serialize tensors
     to the file, or deserialize the tensors.
@@ -421,11 +71,14 @@ class GooseTensorizer:
 
     def __init__(
             self,
-            file_obj: Union[io.BufferedIOBase, io.RawIOBase, str],
-            compress_tensors=False) -> None:
+            file_obj: Union[io.BufferedIOBase, io.RawIOBase, typing.BinaryIO, str],
+            write: bool = False,
+            compress_tensors: bool = False) -> None:
         if isinstance(file_obj, str):
-            self._file = open(file_obj, "rb+")
-            self._file.seek(0)
+            if write:
+                self._file = stream_io.open_stream(file_obj, "ab+")
+            else:
+                self._file = stream_io.open_stream(file_obj, "rb+")
         else:
             self._file = file_obj
         self._tensors = 0
@@ -435,6 +88,7 @@ class GooseTensorizer:
         self.read_bytes = 0
         self._buffer = bytearray(1024 * 1024)
         self._state_dict = None
+        self._idx = 0
         if self.compress_tensors:
             import lz4.frame
             self.lz4_frame = lz4.frame
@@ -470,7 +124,7 @@ class GooseTensorizer:
         else:
             return self._file.tell()
 
-    def finalize(self) -> None:
+    def close(self) -> None:
         """
         Finalizes the serialization with a zero-length field, and closes out
         the file.
@@ -481,7 +135,7 @@ class GooseTensorizer:
         logger.info(f"Tensors completed serializing to {final_sz} bytes")
         if self.compress_tensors:
             compression_ratio = (
-                self.total_tensor_bytes / self.total_compressed_tensor_bytes
+                    self.total_tensor_bytes / self.total_compressed_tensor_bytes
             )
             logger.info(f"Uncomp'd bytes: {self.total_tensor_bytes}")
             logger.info(f"Comp'd bytes: {self.total_compressed_tensor_bytes}")
@@ -598,9 +252,9 @@ class GooseTensorizer:
 
         if self.compress_tensors:
             comp_report = (
-                f" - tensor:[raw: {tensor_raw_sz},"
-                + f" compressed: {tensor_compressed_sz},"
-                + f" ratio: {compression_ratio:.2f}]"
+                    f" - tensor:[raw: {tensor_raw_sz},"
+                    + f" compressed: {tensor_compressed_sz},"
+                    + f" ratio: {compression_ratio:.2f}]"
             )
         else:
             comp_report = ""
@@ -608,6 +262,21 @@ class GooseTensorizer:
             f"{idx}:{typ}:{name} - {str(dtype_bytes)} - "
             f"{tensor.shape} -> {ds_bytes}{comp_report}"
         )
+
+    def write_module(self,
+                     m: torch.nn.Module,
+                     remove_tensors: bool = False) -> None:
+        for module_name, module in m.named_modules():
+            for name, param in module.named_parameters(recurse=False):
+                label = module_name + "." + name
+                self.write_tensor(self._idx, label, TensorType.PARAM, param)
+                if remove_tensors:
+                    setattr(module, name, None)
+            for name, buf in module.named_buffers(recurse=False):
+                label = module_name + "." + name
+                self.write_tensor(self._idx, label, TensorType.BUFFER, buf)
+                setattr(module, name, None)
+            self._idx += 1
 
     def read_tensors(self) -> Iterator[Tuple[int, int, str, numpy.ndarray]]:
         """
@@ -628,13 +297,13 @@ class GooseTensorizer:
                 idx = name_sz + 5
                 name_bytes = headers[5:idx]
                 name: str = name_bytes.decode("utf-8")
-                dtype_len = struct.unpack("<B", headers[idx : idx + 1])[0]
+                dtype_len = struct.unpack("<B", headers[idx: idx + 1])[0]
                 dtype_end = idx + dtype_len + 1
-                dtype = headers[idx + 1 : dtype_end]
+                dtype = headers[idx + 1: dtype_end]
                 # read the shape amount, according to the serialized format. the shape
                 # length is 1 byte after the dtype end.
                 shape_len = struct.unpack(
-                    "<B", headers[dtype_end : dtype_end + 1]
+                    "<B", headers[dtype_end: dtype_end + 1]
                 )[0]
                 # the shape elements are <I so we read 4 bytes. _read_shapes takes in
                 # the header object and the number of elements in the shape.
@@ -643,17 +312,18 @@ class GooseTensorizer:
                 # the shape. so, we need to read 4 * shape_len bytes after the
                 # dtype end + 1 byte for the shape length. sort of convoluted, but it
                 # works.
-                shapelist = self._read_shapes(
-                    headers[dtype_end + 1 : (dtype_end + 1) + (4 * shape_len)],
+                shape_list = self._read_shapes(
+                    headers[dtype_end + 1: (dtype_end + 1) + (4 * shape_len)],
                     shape_len,
                 )
-                datalength = struct.unpack("<q", headers[header_len - 8 :])[0]
-                if datalength > len(self._buffer):
-                    self._buffer = bytearray(datalength)
-                self._file.readinto(memoryview(self._buffer)[:datalength])
+                data_length = struct.unpack("<q", headers[header_len - 8:])[0]
+                if data_length > len(self._buffer):
+                    self._buffer = bytearray(data_length)
+                with memoryview(self._buffer) as mem:
+                    self._file.readinto(mem[:data_length])
                 arr = numpy.ndarray.__new__(
                     numpy.memmap,
-                    shapelist,
+                    shape_list,
                     dtype=dtype,
                     buffer=self._buffer,
                     offset=0,
@@ -664,12 +334,12 @@ class GooseTensorizer:
             return
 
     def state_dict(
-        self,
-        device=get_device(),
-        dtype: [None, str] = None,
+            self,
+            device=utils.get_device(),
+            dtype: [None, str] = None,
     ) -> OrderedDict:
         """
-        Load the tensors in this GooseTensorizer object into a state_dict.
+        Load the tensors in this Tensorizer object into a state_dict.
 
         :param device:
         :param dtype:
@@ -699,26 +369,24 @@ class GooseTensorizer:
 
     def write_state_dict(self, state_dict: Dict):
         """
-        Write the state_dict to the file.
+        Write the state_dict to the file in Tensorizer format.
         """
         idx = 0
         for name, param in state_dict.items():
             self.write_tensor(idx, name, TensorType.STATEDICT, param)
         self.total_tensor_bytes = self._file.tell()
-        self.finalize()
 
     def load_tensors(
-        self,
-        m: torch.nn.Module,
-        device=get_device(),
-        dtype: [None, str] = None,
+            self,
+            m: torch.nn.Module,
+            device=utils.get_device(),
+            dtype: [None, str] = None,
     ) -> int:
         """
         Given `m`, a torch.nn.Module, load the associate tensors in this
-        GooseTensorizer object into the `torch.nn.Module`. Returns the number
-        of tensors loaded into the model.
+        Tensorizer object into the `torch.nn.Module`. Returns the number of tensors
+        loaded into the model.
 
-        Note that this assumes that the `m` is serialized along with this.
         """
         if self._file.closed:
             raise IOError("IO closed, instantiate if you want to load again.")
@@ -752,328 +420,3 @@ class GooseTensorizer:
         self.total_tensor_bytes = self._file.tell()
         self._file.close()
         return tensor_ct
-
-
-
-
-def get_ram_usage_str() -> str:
-    if resource is not None:
-        maxrss_b4 = resource.getrusage(resource.RUSAGE_SELF).ru_maxrss / (
-            1000 * 1000
-        )
-        maxrss_b4_gb = f"{maxrss_b4:0.2f}gb CPU RAM used"
-    else:
-        maxrss_b4_gb = "unknown CPU RAM used"
-    return maxrss_b4_gb
-
-
-def get_vram_usage_str() -> str:
-    if torch.cuda.is_available():
-        gb_gpu = int(
-            torch.cuda.get_device_properties(0).total_memory
-            / (1000 * 1000 * 1000)
-        )
-        return f"{str(gb_gpu)}gb"
-    return "N/A"
-
-
-def get_gpu_name() -> str:
-    if torch.cuda.is_available():
-        return torch.cuda.get_device_name(0)
-    return "N/A"
-
-
-
-
-
-def serialize_model(
-    model: torch.nn.Module,
-    config: Optional[Union[ConfigMixin, AutoConfig, dict]],
-    model_directory: str,
-    model_prefix: str = "model",
-):
-    """
-    Remove the tensors from a PyTorch model, convert them to NumPy
-    arrays and serialize them to GooseTensor format. The stripped
-    model is also serialized to pytorch format.
-
-    Args:
-        model: The model to serialize.
-        config: The model's configuration. This is optional and only
-            required for HuggingFace Transformers models. Diffusers
-            models do not require this.
-        model_directory: The directory to save the serialized model to.
-        model_prefix: The prefix to use for the serialized model files. This
-            is purely optional and it allows for multiple models to be
-            serialized to the same directory. A good example are Stable
-            Diffusion models. Default is "model".
-    """
-
-    os.makedirs(model_directory, exist_ok=True)
-    dir_prefix = f"{model_directory}/{model_prefix}"
-
-    if config is None:
-        config = model
-    if config is not None:
-        if hasattr(config, "to_json_file"):
-            config.to_json_file(f"{dir_prefix}-config.json")
-        if isinstance(config, dict):
-            open(f"{dir_prefix}-config.json", "w").write(
-                json.dumps(config, indent=2)
-            )
-
-    ts = GooseTensorizer(open(f"{dir_prefix}.tensors", "wb"))
-
-    idx = 0
-    for module_name, module in model.named_modules():
-        for name, param in module.named_parameters(recurse=False):
-            v = param.cpu().detach().numpy()
-            ts.write_tensor(idx, module_name + "." + name, TensorType.PARAM, v)
-            setattr(module, name, None)
-        for name, buf in module.named_buffers(recurse=False):
-            v = buf.cpu().detach().numpy()
-            ts.write_tensor(idx, module_name + "." + name, TensorType.BUFFER, v)
-            setattr(module, name, None)
-        idx += 1
-
-    ts.finalize()
-
-
-def load_model(
-    path_uri: str,
-    modelclass: Union[PreTrainedModel, ModelMixin, ConfigMixin] = None,
-    configclass: Optional[Union[ConfigMixin, AutoConfig]] = None,
-    model_prefix: str = "model",
-    device: torch.device = get_device(),
-    dtype: str = None,
-) -> torch.nn.Module:
-    """
-    Given a path prefix, load the model with a custom extension
-
-    Args:
-        path_uri: path to the model. Can be a local path or a URI
-        modelclass: The model class to load the tensors into.
-        configclass: The config class to load the model config into. This must be
-            set if you are loading a model from HuggingFace Transformers.
-        model_prefix: The prefix to use to distinguish between multiple serialized
-            models. The default is "model".
-        dtype: The dtype to load the tensors into. If None, the dtype is inferred from
-            the model.
-    """
-
-    if model_prefix is None:
-        model_prefix = "model"
-
-    if path_uri.startswith("https://") or path_uri.startswith("http://"):
-        config_uri = f"{path_uri}/{model_prefix}-config.json"
-        tensors_uri = f"{path_uri}/{model_prefix}.tensors"
-        if fcntl is not None:
-            # We have fcntl, so we can use the fast CURL-based loader.
-            logger.info("Using CURL for tensor streaming")
-            tensor_loader = lambda: CURLStreamFile(tensors_uri)
-        else:
-            # Fallback to slow requests-based loader.
-            logger.info("Using requests for tensor streaming")
-            tensor_loader = lambda: RequestsStreamFile(tensors_uri)
-    else:
-        tensors_uri = f"{path_uri}/{model_prefix}.tensors"
-        config_uri = "file://" + os.path.join(
-            os.path.dirname(path_uri), f"{model_prefix}-config.json"
-        )
-        if not os.path.exists(config_uri):
-            config_uri = f"{path_uri}/{model_prefix}-config.json"
-        tensor_loader = lambda: open(tensors_uri, "rb")
-
-    logger.info(f"Loading {tensors_uri}, {get_ram_usage_str()}")
-    begin_load = time.time()
-
-    tensor_deserializer = GooseTensorizer(tensor_loader())
-
-    if configclass is not None:
-        try:
-            with tempfile.TemporaryDirectory() as dir:
-                request.urlretrieve(
-                    config_uri, os.path.join(dir, "config.json")
-                )
-                config = configclass.from_pretrained(dir)
-                config.gradient_checkpointing = True
-        except ValueError:
-            config = configclass.from_pretrained(config_uri)
-        model = no_init_or_tensor(
-            lambda: modelclass.from_pretrained(
-                None, config=config, state_dict=OrderedDict()
-            )
-        )
-    else:
-        try:
-            config = json.loads(
-                request.urlopen(config_uri).read().decode("utf-8")
-            )
-        except ValueError:
-            with open(config_uri, "r") as f:
-                config = json.load(f)
-        model = no_init_or_tensor(lambda: modelclass(**config))
-
-    tensor_deserializer.load_tensors(model, device=device, dtype=dtype)
-
-    tensor_load_s = time.time() - begin_load
-    rate_str = convert_bytes(
-        tensor_deserializer.total_bytes_read / tensor_load_s
-    )
-    tensors_sz = convert_bytes(tensor_deserializer.total_bytes_read)
-    logger.info(
-        f"Model tensors loaded in {tensor_load_s:0.2f}s, read "
-        + f"{tensors_sz} @ {rate_str}/s, {get_ram_usage_str()}"
-    )
-
-    return model
-
-
-def df_main(args: argparse.Namespace) -> None:
-    output_prefix = args.output_prefix
-    print("MODEL PATH:", args.input_directory)
-    print("OUTPUT PREFIX:", output_prefix)
-
-    hf_api_token = os.environ.get("HF_API_TOKEN")
-
-    pipeline = StableDiffusionPipeline.from_pretrained(
-        args.input_directory, use_auth_token=hf_api_token
-    )
-
-    logger.info("Serializing model")
-    logger.info("GPU: " + get_gpu_name())
-    logger.info("GPU RAM: " + get_vram_usage_str())
-    logger.info("PYTHON USED RAM: " + get_ram_usage_str())
-
-    serialize_model(
-        pipeline.text_encoder.eval(),
-        pipeline.text_encoder.config,
-        output_prefix,
-        "encoder",
-    )
-    serialize_model(pipeline.vae.eval(), None, output_prefix, "vae")
-    serialize_model(pipeline.unet.eval(), None, output_prefix, "unet")
-
-    pipeline.tokenizer.save_pretrained(output_prefix)
-
-    if args.validate:
-        device = get_device()
-
-        logger.info("Validating serialization")
-        vae = load_model(output_prefix, AutoencoderKL, None, "vae", device)
-        unet = load_model(output_prefix, UNet2DConditionModel, None, "unet")
-        encoder = load_model(
-            output_prefix, CLIPTextModel, CLIPTextConfig, "encoder", device
-        )
-
-        pipeline = StableDiffusionPipeline(
-            text_encoder=encoder,
-            vae=vae,
-            unet=unet,
-            tokenizer=CLIPTokenizer.from_pretrained(
-                args.input_directory, subfolder="tokenizer"
-            ),
-            scheduler=LMSDiscreteScheduler(
-                beta_end=0.012,
-                beta_schedule="scaled_linear",
-                beta_start=0.00085,
-                num_train_timesteps=1000,
-                trained_betas=None,
-            ),
-            safety_checker=None,
-            feature_extractor=None,
-        ).to(device)
-
-        prompt = "a photo of an astronaut riding a horse on mars"
-        with torch.autocast(
-            "cuda" if torch.cuda.is_available() else "cpu"
-        ):  # for some reason device_type needs to be a string instead of an actual device
-            pipeline(prompt).images[0]
-
-
-def hf_main(args):
-    output_prefix = args.output_prefix
-    print("MODEL PATH:", args.input_directory)
-    print("OUTPUT PREFIX:", output_prefix)
-
-    from transformers import AutoModelForCausalLM, AutoConfig, AutoTokenizer
-
-    model_config = AutoConfig.from_pretrained(args.input_directory)
-    model = AutoModelForCausalLM.from_pretrained(
-        args.input_directory, config=model_config, torch_dtype=torch.float16
-    )
-
-    logger.info("Serializing model")
-    logger.info("GPU: " + get_gpu_name())
-    logger.info("GPU RAM: " + get_vram_usage_str())
-    logger.info("PYTHON USED RAM: " + get_ram_usage_str())
-
-    serialize_model(model, model_config, output_prefix)
-
-    tokenizer = AutoTokenizer.from_pretrained(
-        args.input_directory
-    ).save_pretrained(output_prefix)
-
-    if args.validate:
-        device = get_device()
-        logger.info("Validating serialization")
-        model = load_model(
-            output_prefix,
-            AutoModelForCausalLM,
-            AutoConfig,
-            None,
-            device,
-            "float16",
-        ).eval()
-        # test generation
-        tokenizer = AutoTokenizer.from_pretrained(args.input_directory)
-        input_ids = tokenizer.encode(
-            "¡Hola! Encantado de conocerte. hoy voy a", return_tensors="pt"
-        ).to(device)
-        with torch.no_grad():
-            output = model.generate(
-                input_ids, max_new_tokens=50, do_sample=True
-            )
-        logger.info(
-            f"Test Output: {tokenizer.decode(output[0], skip_special_tokens=True)}"
-        )
-
-
-def main():
-    # usage: tensorizer [input-directory] [output-prefix] [model-type]
-
-    parser = argparse.ArgumentParser()
-    parser.add_argument(
-        "input_directory",
-        type=str,
-        help="Path to model directory or HF model ID",
-    )
-    parser.add_argument(
-        "output_prefix", type=str, help="Path to output directory"
-    )
-    parser.add_argument(
-        "--model_type",
-        type=str,
-        choices=["transformers", "diffusers"],
-        required=True,
-        help="Framework used for the model",
-    )
-    parser.add_argument(
-        "--validate",
-        action="store_true",
-        help="Validate serialization by running a test inference",
-    )
-    args = parser.parse_args()
-
-    if args.model_type == "transformers":
-        hf_main(args)
-    elif args.model_type == "diffusers":
-        df_main(args)
-    else:
-        raise ValueError(
-            f"Unknown model type {args.model_type} (transformers or diffusers)"
-        )
-
-
-if __name__ == "__main__":
-    main()
